chore: rewrite examples/oauth scripts in TypeScript
