+++ examples/oauth/send-to-ln-address.ts
@@ -2,7 +2,10 @@ import * as readline from "node:readline/promises";
 import { stdin as input, stdout as output } from "node:process";
 
 import { oauth } from "../../dist/index.module.js";
+import type { oauth as OAuth } from "../../dist/index";
+
 const { auth, Client } = oauth;
+import { LightningAddress } from "alby-tools";
 
 const rl = readline.createInterface({ input, output });
 
@@ -10,13 +13,13 @@ const authClient = new auth.OAuth2User({
   client_id: process.env.CLIENT_ID,
   client_secret: process.env.CLIENT_SECRET,
   callback: "http://localhost:8080",
-  scopes: ["invoices:read", "account:read", "balance:read"],
+  scopes: ["payments:send"],
   token: {
     access_token: undefined,
     refresh_token: undefined,
     expires_at: undefined,
   }, // initialize with existing token
-});
+}) as unknown as OAuth.auth.OAuth2User;
 
 console.log(`Open the following URL and authenticate the app:`);
 console.log(await authClient.generateAuthURL());
@@ -27,13 +30,14 @@ rl.close();
 
 await authClient.requestAccessToken(code);
 console.log(authClient.token);
-const client = new Client(authClient);
+const client = new Client(authClient) as OAuth.Client;
 
-const response = await client.incomingInvoices();
+const ln = new LightningAddress("hello@getalby.com");
+// fetch the LNURL data
+await ln.fetch();
+
+const invoice = await ln.requestInvoice({ satoshi: 1000 });
+
+const response = await client.sendPayment({ invoice: invoice.paymentRequest });
 
 console.log(JSON.stringify(response, null, 2));
-
-if (response[0]) {
-  const invoice = await client.getInvoice(response[0].r_hash_str);
-  console.log(JSON.stringify(invoice, null, 2));
-}
